refactor : 리뷰 작성을 위한 주문 검증 로직 모듈화
